Use decimal for hours and parsing instead of double

+++ TestWinForms/TestWinForms/Models/WorkEntry.cs
@@ -10,6 +10,6 @@ namespace Crotating.Models
     {
         public string Name { get; set; }
         public DateTime Date { get; set; }
-        public double Hours { get; set; }
+        public decimal Hours { get; set; }
     }
 }
